+++ src/services/workspace.service.ts
@@ -1,8 +1,10 @@
+import { Workspace } from "./../models/workspace.model";
+import { getRepository } from "typeorm";
 import { User } from "../models/user.model";
-import { Workspace } from "../models/workspace.model";
 import workspaceRepository from "../repository/workspace.repository";
 import workspaceUserRepository from "../repository/workspaceUser.repository";
 import activityService from "./activity.service";
+import AppError from "../utils/appError.utils";
 
 class WorkspaceService {
   async createWorkspace(workspaceData: Partial<Workspace>, user: User) {
@@ -42,20 +44,37 @@ class WorkspaceService {
   }
 
   async removeWorkspace(workspaceId: string, userId: number) {
+    const isOwnerPresent = await getRepository(Workspace).findOne({
+      where: {
+        id: workspaceId,
+        owner: {
+          id: userId,
+        },
+      },
+    });
+
+    if (!isOwnerPresent) {
+      throw new AppError("Only Owner can delete Workspace", 500);
+    }
+
     await activityService.logWorkspaceActivity(
-      "create-workspace",
+      "delete-workspace",
       userId,
       workspaceId
     );
-
+    // return isOwnerPresent;
     return await workspaceRepository.deleteWorkspace(workspaceId);
   }
 
   async getOwnerWorkspace(ownerId: number) {
     return await workspaceRepository.getOwnerWorkspace(ownerId);
   }
-  async updateWOrkspace(workspaceId: string, name: string) {
-    return await workspaceRepository.updateWorkspace(workspaceId, name);
+  async updateWOrkspace(workspaceId: string, name: string, isPrivate: boolean) {
+    return await workspaceRepository.updateWorkspace(
+      workspaceId,
+      name,
+      isPrivate
+    );
   }
 }
